# Performance optimization: SendAsync: call TrySend within Task