--- a/source/NetCoreServer/SslSession.cs
+++ b/source/NetCoreServer/SslSession.cs
@@ -4,6 +4,7 @@ using System.Net.Security;
 using System.Net.Sockets;
 using System.Text;
 using System.Threading;
+using System.Threading.Tasks;
 
 namespace NetCoreServer
 {
@@ -347,7 +348,7 @@ namespace NetCoreServer
             }
 
             // Try to send the main buffer
-            TrySend();
+            Task.Factory.StartNew(TrySend);
 
             return true;
         }